gate define, middleware design and role wise authorization

+++ app/Http/Controllers/Backend/ModuleController.php
@@ -7,6 +7,7 @@ use Illuminate\Support\Str;
 use Illuminate\Http\Request;
 use App\Http\Controllers\Controller;
 use Brian2694\Toastr\Facades\Toastr;
+use Illuminate\Support\Facades\Gate;
 use App\Http\Requests\ModuleStoreRequest;
 
 class ModuleController extends Controller
@@ -18,6 +19,8 @@ class ModuleController extends Controller
      */
     public function index()
     {
+        Gate::authorize('index-module');
+        //authorize this user to access or not
         $modules = Module::select(['id','module_name','module_slug','updated_at'])->latest('id')->get();
         return view('admin.pages.module.index',compact('modules'));
     }
@@ -29,6 +32,8 @@ class ModuleController extends Controller
      */
     public function create()
     {
+        Gate::authorize('create-module');
+        //authorize this user to access or not
         return view('admin.pages.module.create');
     }
 
@@ -40,6 +45,8 @@ class ModuleController extends Controller
      */
     public function store(ModuleStoreRequest $request)
     {
+        Gate::authorize('create-module');
+        //authorize this user to access or not
         Module::updateOrCreate([
             'module_name' =>$request->module_name,
             'module_slug' =>Str::slug($request->module_name),
@@ -69,6 +76,8 @@ class ModuleController extends Controller
      */
     public function edit($module_slug)
     {
+        Gate::authorize('edit-module');
+        //authorize this user to access or not
         $module = Module::whereModule_slug($module_slug)->select('module_name','module_slug')->first();
         return view('admin.pages.module.edit',compact('module'));
 
@@ -83,6 +92,8 @@ class ModuleController extends Controller
      */
     public function update(ModuleStoreRequest $request, $module_slug)
     {
+        Gate::authorize('update-module');
+        //authorize this user to access or not
         $module = Module::whereModule_slug($module_slug)->first();
 
         $module->update([
@@ -102,7 +113,9 @@ class ModuleController extends Controller
      */
     public function destroy($module_slug)
     {
-         $module = Module::whereModule_slug($module_slug);
+        Gate::authorize('delete-module');
+        //authorize this user to access or not
+        $module = Module::whereModule_slug($module_slug);
          $module->delete();
          Toastr::success('Module Delete Successfully', 'Success',);
          return redirect()->route('admin.module.index');
