+++ app/Http/Controllers/Backend/PermissionController.php
@@ -8,6 +8,7 @@ use Illuminate\Support\Str;
 use Illuminate\Http\Request;
 use App\Http\Controllers\Controller;
 use Brian2694\Toastr\Facades\Toastr;
+use Illuminate\Support\Facades\Gate;
 use App\Http\Requests\PermissionStoreRequest;
 
 class PermissionController extends Controller
@@ -19,6 +20,9 @@ class PermissionController extends Controller
      */
     public function index()
     {
+        Gate::authorize('index-permission');
+        //authorize this user to access or not
+
         $modules = Module::select('id','module_name')->get();
         $permissions = Permission::with(['module:id,module_name,module_slug'])
         ->select(['id','module_id','permission_slug','permission_name','updated_at'])
@@ -45,6 +49,8 @@ class PermissionController extends Controller
      */
     public function store(PermissionStoreRequest $request)
     {
+       Gate::authorize('create-permission');
+        //authorize this user to access or not
         Permission::updateOrCreate([
             'module_id'=> $request->module_id,
             'permission_name'=> $request->permission_name,
@@ -86,6 +92,8 @@ class PermissionController extends Controller
      */
     public function update(PermissionStoreRequest $request, $permission_slug)
     {
+        Gate::authorize('edit-permission');
+        //authorize this user to access or not
         $module = Permission::wherePermission_slug($permission_slug)->first();
         $module->update([
             'module_id'=> $request->module_id,
@@ -105,6 +113,8 @@ class PermissionController extends Controller
      */
     public function destroy($permission_slug)
     {
+        Gate::authorize('delete-permission');
+        //authorize this user to access or not
         $module = Permission::wherePermission_slug($permission_slug);
          $module->delete();
          Toastr::success('Permission Delete Successfully', 'Success',);
